+++ ChangeLog.rst
@@ -1,13 +1,11 @@
 Unreleased
 ----------
+- [#18] yougo plugin の実装
 
 Release Notes - 2017-11-27
 --------------------------
 - [#76] pool_sizeの上限に引っかかりエラー
 - [#90] $cleaning taskのコマンドの出力内容に不要な文言が混ざっている
-
-Release Notes - 2017-11-27
---------------------------
 - [#83] uranai実装
 
 Release Notes - 2017-11-21